+++ src/test/groovy/picocli/groovy/ScriptWithoutImports.groovy
@@ -1,0 +1,10 @@
+package picocli.groovy
+
+@picocli.CommandLine.Command(name = 'cmd', description = 'my description')
+
+@picocli.groovy.PicocliScript
+
+@picocli.CommandLine.Option(names = ["-h", "--help"], usageHelp = true)
+@groovy.transform.Field boolean usageHelpRequested = false
+
+["hi"]